requirements.txt
Flask==0.10.1
Flask-Mail==0.9.1
Flask-RESTful==0.3.1
Flask-SQLAlchemy==2.0
flask-marshmallow==0.4.0
nose==1.3.4
passlib==1.6.2
requests==2.5.1
